requirements.txt
cycler==0.10.0
decorator==4.4.2
filterpy==1.4.5
imageio==2.8.0
joblib==0.14.1
kiwisolver==1.1.0
llvmlite==0.31.0
matplotlib==3.2.1
networkx==2.4
pandas==1.0.3
numpy==1.18.2
torch==1.4.0
opencv_python==4.2.0.32
Kivy==1.11.1
numba==0.45.1
filterpy==1.4.5
torchvision==0.5.0
Pillow==7.1.2
pyparsing==2.4.6
python-dateutil==2.8.1
pytz==2019.3
PyWavelets==1.1.1
scikit-image==0.16.2
scikit_learn==0.23.1
scipy==1.4.1
six==1.14.0
torch==1.4.0
skimage==0.0
git+https://github.com/RomeoSajina/Python-GENEActiv-Reader.git
git+https://github.com/facebookresearch/detectron2.git
pyyaml==5.1
pycocotools>=2.0.1
